Move yargs command registration to separate file
